scripts/pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "timelapse-colorizer-data"
authors = [
    {name = "", email = "danielt@alleninstitute.org"},
]
description = "data prep scripts"
# readme = ""
requires-python = ">=3.10"
keywords = ["one", "two"]
license = {text = "BSD-3-Clause"}
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "aicsimageio[czi]>=4.11.0",
    "aicspylibczi>=3.0.5",
    "numpy",
    "pandas",
    "pillow",
    "scikit-image",
    "lkaccess @ https://artifactory.corp.alleninstitute.org/artifactory/api/pypi/pypi-release-local/lkaccess/1.4.25/lkaccess-1.4.25-py2.py3-none-any.whl#md5=dc51f01c6452c8b69a5e212b9f5b318d",
    "aicsfiles @ https://artifactory.corp.alleninstitute.org/artifactory/api/pypi/pypi-release-local/aicsfiles/5.3.0/aicsfiles-5.3.0.tar.gz#md5=5beeab7fe302f74a7e89f5604f3dbf78",
    # "aicsfiles @ https://artifactory.corp.alleninstitute.org/artifactory/api/pypi/pypi-release-local/aicsfiles/7.0.0/aicsfiles-7.0.0.tar.gz#md5=42e6b5e857cc77c8ec9da6a368571922",
    "nuc-morph-analysis @ git+https://github.com/aics-int/nuc-morph-analysis.git"
]
dynamic = ["version"]

[project.scripts]
my-script = "my_package.module:function"

# ... other project metadata fields as specified in:
#     https://packaging.python.org/en/latest/specifications/declaring-project-metadata/
